requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2023.10.0
PyNaCl>=1.5.0
groq>=0.4.0

# Build tools for Rust extension (optional - graceful fallback if not built)
maturin>=1.0.0
